refactor: AuthenticationService 슬라이싱 테스트로 변경
